services/goldfin-ocr/requirements.txt
ABBYY >= 0.3
PyYAML >= 3.12
boto3 >= 1.5.26
certifi >= 14.05.14
lxml >= 3.8.0
python_dateutil >= 2.5.3
setuptools >= 21.0.0
six >= 1.10
urllib3 >= 1.15.1
